# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Blog CRUD Operation",
        "title": "Blog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/blog_api/v1",
    "paths": {
        "/blog/comment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Add a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blog ID",
                        "name": "blog_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Comment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comment added successfully",
                        "schema": {
                            "$ref": "#/definitions/types.BlogResp"
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error adding comment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Delete a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blog ID",
                        "name": "blog_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comment deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error deleting comment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/blog/comments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get comments of a blog post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get comments of a blog post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blog ID",
                        "name": "blog_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comment IDs",
                        "name": "comment_ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comments fetched successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.CommentResp"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error getting comments",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/blog/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a blog post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Create a blog post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Blog Post Request",
                        "name": "blogPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.BlogPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "blog post created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.BlogResp"
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error creating blog",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/blog/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a blog post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Delete a blog post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blog ID",
                        "name": "blog_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "blog deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error deleting blog",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/blog/get": {
            "get": {
                "description": "Get a blog post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get a blog post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blog ID",
                        "name": "blog_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "blog fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/types.BlogResp"
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error getting blog",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/blog/get/user": {
            "get": {
                "description": "Get all blog posts of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get all blog posts of a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blog IDs",
                        "name": "blog_ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "blogs fetched successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.BlogResp"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error getting blogs",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/blog/getAll": {
            "get": {
                "description": "Get all blog posts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get all blog posts",
                "responses": {
                    "200": {
                        "description": "blogs fetched successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.BlogResp"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid data request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error getting blogs",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.BlogPostRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "content_text": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "photo_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.BlogResp": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CommentResp"
                    }
                },
                "comments_count": {
                    "type": "integer"
                },
                "content_text": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "photo_url": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "reactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ReactionResp"
                    }
                },
                "reactions_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "types.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "types.CommentResp": {
            "type": "object",
            "properties": {
                "blog_post_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "types.ReactionResp": {
            "type": "object",
            "properties": {
                "blog_post_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}